+++ MidiPlayer/src/main/java/org/primesoft/midiplayer/commands/GlobalPlayMidiCommand.java
@@ -50,6 +50,7 @@ import org.primesoft.midiplayer.MusicPlayer;
 import org.primesoft.midiplayer.midiparser.NoteFrame;
 import org.primesoft.midiplayer.midiparser.NoteTrack;
 import org.primesoft.midiplayer.track.GlobalTrack;
+import org.primesoft.midiplayer.utils.CommandUtils;
 
 
 /**
