+++ MidiPlayer/src/main/java/org/primesoft/midiplayer/commands/PlayMidiCommand.java
@@ -55,6 +55,7 @@ import org.primesoft.midiplayer.midiparser.NoteFrame;
 import org.primesoft.midiplayer.midiparser.NoteTrack;
 import org.primesoft.midiplayer.track.BasePlayerTrack;
 import org.primesoft.midiplayer.track.PlayerTrack;
+import org.primesoft.midiplayer.utils.CommandUtils;
 
 import java.util.*;
