+++ MidiPlayer/src/main/java/org/primesoft/midiplayer/commands/StopMidiCommand.java
@@ -48,6 +48,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 import org.primesoft.midiplayer.MusicPlayer;
 import org.primesoft.midiplayer.track.BasePlayerTrack;
+import org.primesoft.midiplayer.utils.CommandUtils;
 
 import java.util.List;
