+++ MidiPlayer/src/main/java/org/primesoft/midiplayer/commands/suggestions/MIDISuggestionProvider.java
@@ -1,4 +1,4 @@
-package org.primesoft.midiplayer.commands;
+package org.primesoft.midiplayer.commands.suggestions;
 
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
